feat: 添加题解帖子创建及查询功能，支持获取题目数和当天提交数

+++ src/main/java/com/zxr/bsoj/model/entity/Posts.java
@@ -37,7 +37,7 @@ public class Posts implements Serializable {
     /**
      * 题目ID（外键，可为空）
      */
-    private Integer problemId;
+    private Long problemId;
 
     /**
      * 编程语言（题解专用），可为空
